refactor: Add conditional check for content length before triggering window
This commit adds a conditional check for the length of the `content` before triggering the window in the `CSPARQLWindow` class. If the `content` has a length greater than 0, the window is triggered and the `RStream` event is emitted. Otherwise, a log message is added to indicate that the window has no data.